Ripped out paypal/bank transfer & intro’d stripe
Going forward, we want to get rid of Paypal because they are about to block access to our account and their systems are very difficult to deal with. Propose that we move forward with Stripe to exclusively capture payments so that we reduce the administrative workload of having to check the club bank account to approve memberships.